+++ src/Voting.Stimmunterlagen.Core/Managers/ContestVotingCardLayoutManager.cs
@@ -55,7 +55,7 @@ public class ContestVotingCardLayoutManager
         _mapper = mapper;
     }
 
-    public async Task SetLayout(Guid contestId, VotingCardType vcType, bool allowCustom, int templateId, VotingCardLayoutDataConfiguration dataConfiguration)
+    public async Task SetLayout(Guid contestId, VotingCardType vcType, bool allowCustom, int templateId, VotingCardLayoutDataConfiguration dataConfiguration, VotingCardColor color)
     {
         var existingLayout = await _contestLayoutRepo.Query()
             .AsTracking()
@@ -74,6 +74,7 @@ public class ContestVotingCardLayoutManager
         existingLayout.AllowCustom = allowCustom;
         existingLayout.TemplateId = templateId;
         existingLayout.DataConfiguration = dataConfiguration;
+        existingLayout.VotingCardColor = color;
 
         await using var transaction = await _dbContext.Database.BeginTransactionAsync(System.Data.IsolationLevel.ReadCommitted);
 
@@ -94,7 +95,10 @@ public class ContestVotingCardLayoutManager
 
             doiLayout.DomainOfInfluenceTemplateId = null;
             doiLayout.OverriddenTemplateId = null;
+            doiLayout.DomainOfInfluenceVotingCardColor = null;
+            doiLayout.OverriddenVotingCardColor = null;
             doiLayout.TemplateId = existingLayout.TemplateId;
+            doiLayout.VotingCardColor = existingLayout.VotingCardColor;
             doiLayout.AllowCustom = existingLayout.AllowCustom;
             doiLayout.DataConfiguration = _mapper.Map<VotingCardLayoutDataConfiguration>(dataConfiguration);
             DataConfigurationValidator.Validate(doiLayout.DataConfiguration, doiLayout.DomainOfInfluence!.StistatMunicipality, contest.IsPoliticalAssembly);
@@ -135,6 +139,7 @@ public class ContestVotingCardLayoutManager
             throw new EntityNotFoundException(nameof(layout.Template), new { contestId, vcType });
         }
 
-        return await _templateManager.GetPdfPreview(null, layout.TemplateId.Value, layout.Contest!, layout.DataConfiguration, cancellationToken: ct);
+        var doi = DummyDoiBuilder.GetDummyDomainOfInfluence(_auth.Tenant.Id, layout.PrintData, layout.VotingCardColor);
+        return await _templateManager.GetPdfPreview(null, layout.TemplateId.Value, layout.Contest!, layout.DataConfiguration, doi, cancellationToken: ct);
     }
 }