+++ src/Voting.Stimmunterlagen.Core/Managers/DomainOfInfluenceVotingCardLayoutManager.cs
@@ -57,7 +57,7 @@ public class DomainOfInfluenceVotingCardLayoutManager
         _stepStateRepo = stepStateRepo;
     }
 
-    public async Task SetLayout(Guid doiId, VotingCardType vcType, bool allowCustom, int? templateId, VotingCardLayoutDataConfiguration dataConfiguration)
+    public async Task SetLayout(Guid doiId, VotingCardType vcType, bool allowCustom, int? templateId, VotingCardLayoutDataConfiguration dataConfiguration, VotingCardColor? color)
     {
         var existingLayout = await _doiLayoutRepo.Query()
             .AsTracking()
@@ -79,6 +79,8 @@ public class DomainOfInfluenceVotingCardLayoutManager
 
         var oldEffectiveTemplateId = existingLayout.EffectiveTemplateId;
         existingLayout.OverriddenTemplateId = null;
+        existingLayout.OverriddenVotingCardColor = null;
+        existingLayout.DomainOfInfluenceVotingCardColor = color;
         existingLayout.AllowCustom = allowCustom;
         DataConfigurationValidator.Validate(dataConfiguration, existingLayout.DomainOfInfluence!.StistatMunicipality, existingLayout.DomainOfInfluence!.Contest!.IsPoliticalAssembly);
         existingLayout.DataConfiguration = dataConfiguration;
@@ -105,7 +107,7 @@ public class DomainOfInfluenceVotingCardLayoutManager
         await _doiLayoutRepo.SaveChanges();
     }
 
-    public async Task SetOverriddenLayout(Guid doiId, VotingCardType vcType, int? templateId, VotingCardLayoutDataConfiguration dataConfiguration)
+    public async Task SetOverriddenLayout(Guid doiId, VotingCardType vcType, int? templateId, VotingCardLayoutDataConfiguration dataConfiguration, VotingCardColor color)
     {
         await using var transaction = await _dbContext.Database.BeginTransactionAsync(IsolationLevel.ReadCommitted);
 
@@ -149,6 +151,8 @@ public class DomainOfInfluenceVotingCardLayoutManager
             template = await _templateManager.GetOrCreateTemplate(templateId.Value);
             existingLayout.OverriddenTemplateId = templateId.Value;
         }
+
+        existingLayout.OverriddenVotingCardColor = color;
 
         SyncTemplateFields(existingLayout, template, true);
         await _doiLayoutRepo.SaveChanges();