+++ src/Voting.Stimmunterlagen.Data/Migrations/20260826140617_VotingCardColorUpdate.Designer.cs
@@ -2,6 +2,7 @@
 using System;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 using Voting.Stimmunterlagen.Data;
@@ -11,9 +12,11 @@ using Voting.Stimmunterlagen.Data;
 namespace Voting.Stimmunterlagen.Data.Migrations
 {
     [DbContext(typeof(DataContext))]
-    partial class DataContextModelSnapshot : ModelSnapshot
+    [Migration("20260826140617_VotingCardColorUpdate")]
+    partial class VotingCardColorUpdate
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -539,9 +542,6 @@ namespace Voting.Stimmunterlagen.Data.Migrations
                     b.Property<int?>("TemplateId")
                         .HasColumnType("integer");
 
-                    b.Property<int>("VotingCardColor")
-                        .HasColumnType("integer");
-
                     b.Property<int>("VotingCardType")
                         .HasColumnType("integer");
 
@@ -823,19 +823,10 @@ namespace Voting.Stimmunterlagen.Data.Migrations
                     b.Property<int?>("DomainOfInfluenceTemplateId")
                         .HasColumnType("integer");
 
-                    b.Property<int?>("DomainOfInfluenceVotingCardColor")
-                        .HasColumnType("integer");
-
                     b.Property<int?>("OverriddenTemplateId")
                         .HasColumnType("integer");
 
-                    b.Property<int?>("OverriddenVotingCardColor")
-                        .HasColumnType("integer");
-
                     b.Property<int?>("TemplateId")
-                        .HasColumnType("integer");
-
-                    b.Property<int>("VotingCardColor")
                         .HasColumnType("integer");
 
                     b.Property<int>("VotingCardType")