+++ src/Voting.Stimmunterlagen/MappingProfiles/EnumProfile.cs
@@ -28,14 +28,16 @@ public class EnumProfile : Profile
         CreateEnumMap<VotingCardSort, ProtoModels.VotingCardSort>();
         CreateEnumMap<PrintJobState, ProtoModels.PrintJobState>();
         CreateEnumMap<VotingCardGeneratorJobState, ProtoModels.VotingCardGeneratorJobState>();
+        CreateEnumMap<VotingCardColor, ProtoModels.VotingCardColor>();
 
         // explicitly map deprecated values to default value.
         CreateMap<Abraxas.Voting.Basis.Shared.V1.VotingCardColor, VotingCardColor>()
             .ConvertUsingEnumMapping(opt => opt
                 .MapByValue()
                 .MapValue(Abraxas.Voting.Basis.Shared.V1.VotingCardColor.Chamois, VotingCardColor.Unspecified)
-                .MapValue(Abraxas.Voting.Basis.Shared.V1.VotingCardColor.Gold, VotingCardColor.Unspecified))
-            .ReverseMap();
+                .MapValue(Abraxas.Voting.Basis.Shared.V1.VotingCardColor.Gold, VotingCardColor.Unspecified)
+                .MapValue(Abraxas.Voting.Basis.Shared.V1.VotingCardColor.Grey, VotingCardColor.Purple)
+                .MapValue(Abraxas.Voting.Basis.Shared.V1.VotingCardColor.Red, VotingCardColor.Orange));
     }
 
     private void CreateEnumMap<T1, T2>()