+++ test/Voting.Stimmunterlagen.IntegrationTest/DomainOfInfluenceVotingCardLayoutTests/SetDomainOfInfluenceVotingCardLayoutTest.cs
@@ -16,6 +16,7 @@ using Voting.Stimmunterlagen.IntegrationTest.MockData;
 using Voting.Stimmunterlagen.Proto.V1;
 using Voting.Stimmunterlagen.Proto.V1.Requests;
 using Xunit;
+using VotingCardColor = Voting.Stimmunterlagen.Proto.V1.Models.VotingCardColor;
 using VotingCardType = Voting.Stimmunterlagen.Proto.V1.Models.VotingCardType;
 
 namespace Voting.Stimmunterlagen.IntegrationTest.DomainOfInfluenceVotingCardLayoutTests;
@@ -52,6 +53,7 @@ public class SetDomainOfInfluenceVotingCardLayoutTest :
             {
                 IncludeReligion = true,
             },
+            Color = VotingCardColor.Yellow,
         });
 
         var layout = await RunOnDb(db => db.DomainOfInfluenceVotingCardLayouts
@@ -61,8 +63,12 @@ public class SetDomainOfInfluenceVotingCardLayoutTest :
         layout.AllowCustom.Should().BeTrue();
         layout.TemplateId.Should().Be(DmDocServiceMock.TemplateSwiss.Id);
         layout.DomainOfInfluenceTemplateId.Should().Be(DmDocServiceMock.TemplateOthers2.Id);
-        layout.EffectiveTemplateId.Should().Be(DmDocServiceMock.TemplateOthers2.Id);
         layout.OverriddenTemplateId.Should().BeNull();
+        layout.EffectiveTemplateId.Should().Be(DmDocServiceMock.TemplateOthers2.Id);
+        layout.VotingCardColor.Should().Be(Data.Models.VotingCardColor.Green);
+        layout.DomainOfInfluenceVotingCardColor.Should().Be(Data.Models.VotingCardColor.Yellow);
+        layout.OverriddenVotingCardColor.Should().BeNull();
+        layout.EffectiveVotingCardColor.Should().Be(Data.Models.VotingCardColor.Yellow);
     }
 
     [Fact]
@@ -75,6 +81,7 @@ public class SetDomainOfInfluenceVotingCardLayoutTest :
             VotingCardType = VotingCardType.Swiss,
             DomainOfInfluenceId = DomainOfInfluenceMockData.ContestBundFutureGemeindeArneggId,
             DataConfiguration = new(),
+            Color = VotingCardColor.Yellow,
         });
 
         var layout = await RunOnDb(db => db.DomainOfInfluenceVotingCardLayouts
@@ -133,6 +140,7 @@ public class SetDomainOfInfluenceVotingCardLayoutTest :
             {
                 IncludeIsHouseholder = true,
             },
+            Color = VotingCardColor.Yellow,
         });
         await AbraxasElectionAdminClient.SetLayoutAsync(new SetDomainOfInfluenceVotingCardLayoutRequest
         {
@@ -152,6 +160,10 @@ public class SetDomainOfInfluenceVotingCardLayoutTest :
         layout.EffectiveTemplateId.Should().Be(DmDocServiceMock.TemplateSwiss.Id);
         layout.OverriddenTemplateId.Should().BeNull();
         layout.DataConfiguration.IncludeIsHouseholder.Should().BeFalse();
+        layout.VotingCardColor.Should().Be(Data.Models.VotingCardColor.Green);
+        layout.DomainOfInfluenceVotingCardColor.Should().Be(Data.Models.VotingCardColor.Unspecified);
+        layout.OverriddenVotingCardColor.Should().BeNull();
+        layout.EffectiveVotingCardColor.Should().Be(Data.Models.VotingCardColor.Unspecified);
     }
 
     [Fact]
@@ -171,6 +183,7 @@ public class SetDomainOfInfluenceVotingCardLayoutTest :
                 VotingCardType = VotingCardType.Swiss,
                 DomainOfInfluenceId = doiGuid.ToString(),
                 DataConfiguration = new(),
+                Color = VotingCardColor.Unspecified,
             }),
             StatusCode.NotFound);
     }
@@ -186,6 +199,7 @@ public class SetDomainOfInfluenceVotingCardLayoutTest :
                 VotingCardType = VotingCardType.Swiss,
                 DomainOfInfluenceId = DomainOfInfluenceMockData.ContestBundFutureGemeindeArneggId,
                 DataConfiguration = new(),
+                Color = VotingCardColor.Unspecified,
             }),
             StatusCode.NotFound);
     }
@@ -201,6 +215,7 @@ public class SetDomainOfInfluenceVotingCardLayoutTest :
                 VotingCardType = VotingCardType.Swiss,
                 DomainOfInfluenceId = DomainOfInfluenceMockData.ContestBundFutureApprovedGemeindeArneggId,
                 DataConfiguration = new(),
+                Color = VotingCardColor.Unspecified,
             }),
             StatusCode.NotFound);
     }
@@ -216,6 +231,7 @@ public class SetDomainOfInfluenceVotingCardLayoutTest :
                 VotingCardType = VotingCardType.Swiss,
                 DomainOfInfluenceId = DomainOfInfluenceMockData.ContestBundArchivedNotApprovedGemeindeArneggId,
                 DataConfiguration = new(),
+                Color = VotingCardColor.Unspecified,
             }),
             StatusCode.NotFound);
     }